Unshare the PID namespace of the 'run' child process
I've also introduced a 'grandchild' process which is allocated a PID in
this new PID namespace. Although it isn't strictly needed at this point,
I'm going to need it when I want to mount a separate proc filesystem for
the bash shell. Also it makes the output more comparable with lizrice's
Go version and easier to see that the new PID is in the new namespace.

    root@ubuntu-xenial:/vagrant# ./container run /bin/bash
    run: Running ["/bin/bash"] as PID 14016
    child: Running ["/bin/bash"] as PID 14018
    grandchild: Running ["/bin/bash"] as PID 1
    root@container:/vagrant# echo $$
    4
    root@container:/vagrant# exit
    root@ubuntu-xenial:/vagrant#